Consider the following (legal) sequence:   automaton is created   manager->destroy () is called   parent receives automaton_created
The old code moved the automaton from the create recv queue to the destroy send queue when destroy () was called.
Consequently, the automaton_created_effect couldn't find the automaton in the recv queue.

To fix this, I created another queue, create_done, for automata that are fully created.
An automaton can only move from the destroy_send to destroy_receive queue if the automaton is in the create_done queue.